requirements.txt
# Runtime dependencies for FAST-Calib Python Standalone
# Tested with Python 3.10.
#
# Install with:
#   pip install -r requirements.txt
#
# Linux note:
#   Tkinter is a system package, not a pip package. Install `python3-tk`
#   if you want to use `gui_app.py`.
#
numpy==1.26.4
opencv-python==4.11.0.86
open3d==0.19.0
scipy==1.15.1
scikit-learn==1.5.2
PyYAML==6.0.2
plotly==6.5.0
dash==3.3.0
